pyproject.toml
[build-system]
requires = ["setuptools>=72", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "twingate-log-shipper"
version = "0.1.0"
description = "Ship Twingate connector analytics to S3-compatible storage"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.12"
authors = [{ name = "Twingate Solutions" }]
keywords = ["twingate", "log-shipper", "analytics", "s3"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Logging",
]
dependencies = [
    "aiobotocore>=2.13.0",
    "pydantic>=2.7.0,<3.0",
    "pydantic-settings>=2.3.0,<3.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
journald = ["systemd-python>=235"]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "moto[s3,server]>=5.0.0",
    "boto3>=1.34.0",      # used in test_integration.py to verify S3 uploads via sync client
    "werkzeug>=3.0.0",    # moto server mode dependency (Flask WSGI runner in integration tests)
    "jsonschema>=4.22.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-jsonschema>=4.22.0",
]

[project.scripts]
twingate-log-shipper = "twingate_log_shipper.main:run"

[tool.setuptools.packages.find]
where = ["src"]

# ── Ruff ──────────────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "RUF", # ruff-specific
]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.format]
quote-style = "double"

# ── Mypy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "aiobotocore.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "botocore.*"
ignore_missing_imports = true

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
